Added map tab IOS ONLY
Removed feedScreen since we are using HomeScreen instead. Nothing was in FeedScreen. MapScreen is a tab at the bottom,  and is centered at UCSD. I think it has to be its own page as a requirement. It uses apple maps for now. We can switch to google maps when we set up
DOES NOT WORK FOR ANDROID YET